Introduce wp-utils and move snakToCameCase functions
