+++ packages/wp-dashboard/package.json
@@ -38,6 +38,7 @@
     "@googleforcreators/story-editor": "*",
     "@googleforcreators/tracking": "*",
     "@googleforcreators/url": "*",
+    "@web-stories-wp/wp-utils": "*",
     "@wordpress/api-fetch": "^6.0.1",
     "@wordpress/dom-ready": "^3.3.1",
     "flagged": "^2.0.1",
