+++ packages/wp-dashboard/src/api/user.js
@@ -19,9 +19,9 @@
 import apiFetch from '@wordpress/api-fetch';
 
 /**
- * Internal dependencies
+ * External dependencies
  */
-import { snakeToCamelCaseObjectKeys } from './utils';
+import { snakeToCamelCaseObjectKeys } from '@web-stories-wp/wp-utils';
 
 /**
  * Get user.
