+++ packages/wp-dashboard/src/api/utils/index.js
@@ -14,4 +14,3 @@
  * limitations under the License.
  */
 export { default as reshapeStoryObject } from './reshapeStoryObject';
-export * from './snakeToCamelCase';
