+++ packages/wp-story-editor/package.json
@@ -36,6 +36,7 @@
     "@googleforcreators/story-editor": "*",
     "@googleforcreators/tracking": "*",
     "@googleforcreators/url": "*",
+    "@web-stories-wp/wp-utils": "*",
     "@wordpress/api-fetch": "^6.0.1",
     "@wordpress/dom-ready": "^3.3.1",
     "flagged": "^2.0.1",
