+++ packages/wp-story-editor/src/api/taxonomy.js
@@ -18,15 +18,12 @@
  * External dependencies
  */
 import { addQueryArgs } from '@googleforcreators/url';
+import { snakeToCamelCaseObjectKeys } from '@web-stories-wp/wp-utils';
 
 /**
  * WordPress dependencies
  */
 import apiFetch from '@wordpress/api-fetch';
-/**
- * Internal dependencies
- */
-import { snakeToCamelCaseObjectKeys } from './utils';
 
 /**
  * Get all taxonomies.
