+++ packages/wp-utils/package.json
@@ -1,0 +1,30 @@
+{
+    "name": "@web-stories-wp/wp-utils",
+    "description": "Utilities for WordPress integration packages",
+    "private": true,
+    "author": {
+        "name": "Google",
+        "url": "https://creators.google/"
+    },
+    "license": "Apache-2.0",
+    "keywords": [
+        "web stories",
+        "wordpress"
+    ],
+    "homepage": "https://github.com/GoogleForCreators/web-stories-wp/blob/main/packages/wp-utils/README.md",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/GoogleForCreators/web-stories-wp.git",
+        "directory": "packages/wp-utils"
+    },
+    "bugs": {
+        "url": "https://github.com/googleforcreators/web-stories-wp/issues"
+    },
+    "engines": {
+        "node": ">= 12 || >= 14 || >= 16",
+        "npm": ">= 7.3"
+    },
+    "type": "module",
+    "main": "./src/index.js",
+    "dependencies": {}
+}
